DB initialization integrated into backend
