+++ .gitignore
@@ -1,8 +1,7 @@
 venv/
-backend/backend/__pycache__/
-SummaryMeasures/__pycache__/
-backend/summary_measures/__pycache__/
 backend/summary_measures/migrations/
-database/data/test_data/
+backend/db_connector/migrations/
+backend/db_connector/data/test_data/
 .vscode
 .env
+**/__pycache__/
