+++ backend/db_connector/models.py
@@ -6,7 +6,7 @@ class LightCycleColony(models.Model):
 
 class LightCycleTest(models.Model):
     LightCycleTest_ID = models.IntegerField(primary_key=True)          
-    LightCycleTestDesc = models.CharField(max_length=200) 
+    LightCycleTestDesc = models.CharField(max_length=200,null=True) 
 
 class ArenaType(models.Model):
     ArenaType_ID = models.IntegerField(primary_key=True)          
@@ -52,28 +52,28 @@ class Treatment(models.Model):
     Treatment_ID = models.IntegerField(primary_key=True)           
     SurgeryManipulation_ID = models.ForeignKey(SurgeryManipulation, on_delete=models.CASCADE, db_column='SurgeryManipulation_ID')
     SurgeryOutcome_ID = models.ForeignKey(SurgeryOutcome, on_delete=models.CASCADE, db_column='SurgeryOutcome_ID')        
-    DrugRx_Drug1 = models.CharField(max_length=20)   
-    DrugRx_Dose1 = models.CharField(max_length=100)  
-    DrugRx_Drug2 = models.CharField(max_length=20)   
-    DrugRx_Dose2 = models.CharField(max_length=100)  
-    DrugRx_Drug3 = models.CharField(max_length=20)   
-    DrugRx_Dose3 = models.CharField(max_length=100)  
+    DrugRx_Drug1 = models.CharField(max_length=20,null=True)   
+    DrugRx_Dose1 = models.CharField(max_length=100,null=True)  
+    DrugRx_Drug2 = models.CharField(max_length=20,null=True)   
+    DrugRx_Dose2 = models.CharField(max_length=100,null=True)  
+    DrugRx_Drug3 = models.CharField(max_length=20,null=True)   
+    DrugRx_Dose3 = models.CharField(max_length=100,null=True)  
 
 class Trial(models.Model):
     Trial_ID = models.IntegerField(primary_key=True)          
     DateAndTime = models.CharField(max_length=100) 
-    AnimalWeight = models.IntegerField()          
+    AnimalWeight = models.IntegerField(null=True)          
     InjectionNumber = models.IntegerField()          
-    OFTestNumber = models.IntegerField()          
+    OFTestNumber = models.IntegerField(null=True)          
     DrugRxNumber = models.IntegerField()          
     Experimenter = models.CharField(max_length=100) 
     Duration = models.IntegerField()          
     FallsDuringTest = models.IntegerField()          
-    Notes = models.CharField(max_length=2000)
-    Trackfile = models.CharField(max_length=200) 
-    Pathplot = models.CharField(max_length=200) 
-    Video = models.CharField(max_length=200) 
-    Video_ID = models.CharField(max_length=100)
+    Notes = models.CharField(max_length=2000,null=True)
+    Trackfile = models.CharField(max_length=200,null=True) 
+    Pathplot = models.CharField(max_length=200,null=True) 
+    Video = models.CharField(max_length=200,null=True) 
+    Video_ID = models.CharField(max_length=100,null=True)
     EventType_ID = models.ForeignKey(EventType, on_delete=models.CASCADE, db_column='EventType_ID')
     Animal_ID = models.ForeignKey(Animal, on_delete=models.CASCADE, db_column='Animal_ID')
     Apparatus_ID = models.ForeignKey(Apparatus, on_delete=models.CASCADE, db_column='Apparatus_ID')
@@ -116,10 +116,10 @@ class TimeSeries(models.Model):
     ID = models.AutoField(primary_key=True)
     Trial_ID = models.ForeignKey(Trial, on_delete=models.CASCADE, db_column='Trial_ID')
     Sample_ID = models.IntegerField() 
-    T = models.DecimalField(max_digits=10, decimal_places=2)
-    X = models.DecimalField(max_digits=10, decimal_places=2)
-    Y = models.DecimalField(max_digits=10, decimal_places=2)
-    X_S = models.DecimalField(max_digits=10, decimal_places=2)
-    Y_S = models.DecimalField(max_digits=10, decimal_places=2)
-    V_S = models.DecimalField(max_digits=10, decimal_places=2)
+    T = models.DecimalField(max_digits=10, decimal_places=2,null=True)
+    X = models.DecimalField(max_digits=10, decimal_places=2,null=True)
+    Y = models.DecimalField(max_digits=10, decimal_places=2,null=True)
+    X_S = models.DecimalField(max_digits=10, decimal_places=2,null=True)
+    Y_S = models.DecimalField(max_digits=10, decimal_places=2,null=True)
+    V_S = models.DecimalField(max_digits=10, decimal_places=2,null=True)
     MovementType_S = models.IntegerField()
